refactor(MPT-21793): extract AgreementSubscriptionsSyncer from AgreementSyncer
Split syncer.py into agreement_syncer.py and agreement_subscription_syncer.py,
each with a focused responsibility. Introduced base.py to hold shared types and
avoid circular imports. Added 100% coverage test file for the new syncer class.
