+++ pyproject.toml
@@ -133,7 +133,6 @@ per-file-ignores = [
   "swo_aws_extension/swo/cloud_orchestrator/client.py: WPS214",
   "swo_aws_extension/swo/rql/query_builder.py: WPS110 WPS111 WPS115 WPS210 WPS211 WPS214 WPS221 WPS231 WPS237 WPS478",
   "swo_aws_extension/swo/finops/client.py: WPS231 WPS505 WPS111 WPS121 WPS122 WPS214 WPS420 WPS430 WPS432 WPS458",
-  "swo_aws_extension/swo/mpt/sync/syncer.py: WPS202 WPS204 WPS210 WPS213",
   "swo_aws_extension/flows/jobs/invitations_report_creator.py: WPS214",
   "tests/aws/test_config.py: WPS202",
   "tests/aws/test_errors.py: WPS202",
@@ -143,7 +142,6 @@ per-file-ignores = [
   "tests/swo/ccp/test_client.py: WPS202 WPS211",
   "tests/aws/test_client.py: WPS432 WPS202 WPS204",
   "tests/test_initializer.py: WPS111 WPS213 WPS218 WPS230 WPS431",
-  "tests/swo/mpt/sync/conftest.py: WPS430 WPS202 WPS118",
   "tests/swo/rql/test_query_builder.py: AAA01 AAA05 WPS110 WPS111 WPS122 WPS202 WPS204 WPS210 WPS218 WPS221 WPS347 WPS420 WPS431 WPS432 WPS473 WPS604",
   "tests/processor/querying/test_aws_customer_roles.py: WPS118 WPS202 WPS211",
   "tests/flows/steps/crm_tickets/test_ticket_manager.py: WPS118 WPS211",
